+++ preset/interceptors/index.js
@@ -1,0 +1,2 @@
+export { default as commonCancelRequest } from './commonCancelRequest.js'
+export { default as commonThrottleRequest } from './commonThrottleRequest.js'