+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.4
+* change: preset/interceptors as a folder
+* fix: `commonThrottleRequest` not work correctly when request error
+* fix: `commonThrottleRequest` set notThrottle=true not work
 ## 1.2.3
 * add: interceptors preset `commonThrottleRequest`
 ## 1.2.2